+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/AllyChemtechDragonKillAnimation.java
@@ -2,7 +2,7 @@ package com.bonepl.chromaleague.hud.animations;
 
 import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
-public class AllyChemtechDragonKillAnimation extends DragonKillBlinkingAnimation {
+public class AllyChemtechDragonKillAnimation extends AllyDragonKillBlinkingAnimation {
     public AllyChemtechDragonKillAnimation() {
         super(DragonType.CHEMTECH);
     }
