+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/AllyCloudDragonKillAnimation.java
@@ -2,7 +2,7 @@ package com.bonepl.chromaleague.hud.animations;
 
 import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
-public class AllyCloudDragonKillAnimation extends DragonKillBlinkingAnimation {
+public class AllyCloudDragonKillAnimation extends AllyDragonKillBlinkingAnimation {
     public AllyCloudDragonKillAnimation() {
         super(DragonType.CLOUD);
     }
