+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/AllyHextechDragonKillAnimation.java
@@ -2,7 +2,7 @@ package com.bonepl.chromaleague.hud.animations;
 
 import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
-public class AllyHextechDragonKillAnimation extends DragonKillBlinkingAnimation {
+public class AllyHextechDragonKillAnimation extends AllyDragonKillBlinkingAnimation {
     public AllyHextechDragonKillAnimation() {
         super(DragonType.HEXTECH);
     }
