+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/AllyInfernalDragonKillAnimation.java
@@ -2,7 +2,7 @@ package com.bonepl.chromaleague.hud.animations;
 
 import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
-public class AllyInfernalDragonKillAnimation extends DragonKillBlinkingAnimation {
+public class AllyInfernalDragonKillAnimation extends AllyDragonKillBlinkingAnimation {
     public AllyInfernalDragonKillAnimation() {
         super(DragonType.INFERNAL);
     }
