+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/AllyMountainDragonKillAnimation.java
@@ -2,7 +2,7 @@ package com.bonepl.chromaleague.hud.animations;
 
 import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
-public class AllyMountainDragonKillAnimation extends DragonKillBlinkingAnimation {
+public class AllyMountainDragonKillAnimation extends AllyDragonKillBlinkingAnimation {
     public AllyMountainDragonKillAnimation() {
         super(DragonType.MOUNTAIN);
     }
