+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/AllyOceanDragonKillAnimation.java
@@ -2,7 +2,7 @@ package com.bonepl.chromaleague.hud.animations;
 
 import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
-public class AllyOceanDragonKillAnimation extends DragonKillBlinkingAnimation {
+public class AllyOceanDragonKillAnimation extends AllyDragonKillBlinkingAnimation {
     public AllyOceanDragonKillAnimation() {
         super(DragonType.OCEAN);
     }
