+++ chroma-league/src/main/java/com/bonepl/chromaleague/hud/animations/EnemyDragonKillBlinkingAnimation.java
@@ -4,7 +4,7 @@ import com.bonepl.chromaleague.rest.eventdata.DragonType;
 
 import static com.bonepl.chromaleague.hud.PredefinedKeySets.BLACKWIDOW_FUNCTIONAL;
 
-public abstract class EnemyDragonKillBlinkingAnimation extends DragonKillBlinkingAnimation {
+public abstract class EnemyDragonKillBlinkingAnimation extends AllyDragonKillBlinkingAnimation {
 
     public EnemyDragonKillBlinkingAnimation(DragonType dragonType) {
         super(BLACKWIDOW_FUNCTIONAL, dragonType);
